refactor: 把代码抽离到公共的mod.rs内
